Export company search DTOs from response modules
