+++ application/dtos/company_search_result_dto.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import List
+
+
+@dataclass(frozen=True)
+class CompanySearchResultDTO:
+    company_name: str
+    trading_name: str | None
+    tickers: List[str] = field(default_factory=list)
+    sector: str | None = None
+    subsector: str | None = None
+    segment: str | None = None
+    market: str | None = None
+    institution_common: str | None = None
+    institution_preferred: str | None = None
